src/translate_api/requirements.txt
googletrans==2.1.3
virtualenv
flask-restplus==0.9.2
flask-restful-swagger-2==0.35
Flask-SQLAlchemy==2.1
gunicorn==18.0